Feature/position

See merge request pirimid/ptp/ptp-backend!14

+++ src/main/java/com/pirimidtech/ptp/entity/Position.java
@@ -5,7 +5,16 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.OneToOne;
+import javax.persistence.Table;
 import java.util.UUID;
 
 @Getter
@@ -19,7 +28,7 @@ public class Position {
     @GeneratedValue
     private UUID id;
 
-    private int volume;
+    private float volume;
 
     private float price;
